More Input Options
This is my first extension for TurboWarp. (I don't know how to do it in Scratch, but I love Scratch.) This extension makes it convenient for us to use TurboWarp. (Or you can use it with Scratch—I'm totally fine with that.) By the way, I am Taiwanese, and I am working hard to improve my English.